# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Go File Explorer API
  version: 1.0.0
  description: |
    REST API para exploración y gestión de archivos con autenticación JWT, roles y operaciones asíncronas.

    Roles:
    - viewer: lectura
    - editor: lectura + operaciones de escritura
    - admin: todo + auditoría + registro de usuarios
servers:
  - url: http://localhost:8080
    description: Local
...
paths:
...
  /api/v1/files:
    get:
      tags: [Explorer]
      summary: Listado de directorio
      description: "Rol requerido: viewer/editor/admin"
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: path
          schema: { type: string, example: / }
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - in: query
          name: sort
          schema: { type: string, enum: [name, size, modified_at, type], default: name }
        - in: query
          name: order
          schema: { type: string, enum: [asc, desc], default: asc }
      responses:
        '200':
          description: Listado paginado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DirectoryListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

    delete:
      tags: [Operations]
      summary: Soft-delete de rutas
      description: "Rol requerido: editor/admin"
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteRequest'
      responses:
        '200':
          description: Resultado por ruta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponseEnvelope'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequestError:
      description: Solicitud inválida
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          examples:
            invalidBody: { $ref: '#/components/examples/BadRequest' }
    UnauthorizedError:
      description: Token inválido o ausente
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          examples:
            unauthorized: { $ref: '#/components/examples/Unauthorized' }
    ForbiddenError:
      description: Rol insuficiente
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          examples:
            forbidden: { $ref: '#/components/examples/Forbidden' }
...
  examples:
    BadRequest:
      value:
        success: false
        error:
          code: BAD_REQUEST
          message: invalid JSON body
    Unauthorized:
      value:
        success: false
        error:
          code: UNAUTHORIZED
          message: invalid or expired token
    Forbidden:
      value:
        success: false
        error:
          code: FORBIDDEN
          message: insufficient permissions
...
  schemas:
    APIError:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: string }
      required: [code, message]
    Meta:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        total_pages: { type: integer }
      required: [page, limit, total, total_pages]
    ErrorEnvelope:
      type: object
      properties:
        success: { type: boolean, enum: [false] }
        error: { $ref: '#/components/schemas/APIError' }
      required: [success, error]
...
    FileItem:
      type: object
      properties:
        name: { type: string }
        path: { type: string }
        type: { type: string, enum: [file, directory] }
        size: { type: integer, format: int64 }
        size_human: { type: string }
        mime_type: { type: string }
        extension: { type: string }
        preview_url: { type: string }
        thumbnail_url: { type: string }
        is_image: { type: boolean }
        modified_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
        match_context: { type: string }
        permissions: { type: string }
        item_count: { type: integer }
      required: [name, path, type, size, modified_at, created_at, permissions]
    DirectoryListData:
      type: object
      properties:
        current_path: { type: string }
        parent_path: { type: string }
        items:
          type: array
          items: { $ref: '#/components/schemas/FileItem' }
      required: [current_path, parent_path, items]
    DirectoryListResponse:
      type: object
      properties:
        success: { type: boolean, enum: [true] }
        data: { $ref: '#/components/schemas/DirectoryListData' }
        meta: { $ref: '#/components/schemas/Meta' }
      required: [success, data, meta]
...
    DeleteRequest:
      type: object
      properties:
        paths:
          type: array
          items: { type: string }
      required: [paths]
    DeleteFailure:
      type: object
      properties:
        path: { type: string }
        reason: { type: string }
      required: [path, reason]
    DeleteResponseData:
      type: object
      properties:
        deleted:
          type: array
          items: { type: string }
        failed:
          type: array
          items: { $ref: '#/components/schemas/DeleteFailure' }
      required: [deleted, failed]
    DeleteResponseEnvelope:
      type: object
      properties:
        success: { type: boolean, enum: [true] }
        data: { $ref: '#/components/schemas/DeleteResponseData' }
      required: [success, data]
...
security:
  - BearerAuth: []